Add my own setting.
New keycaps are not shine-through :(, so depend on the LED of the keys
to know which layer is on is not possible. This patch make the indicator
LED to be used in full range, theorically it will support showing the
layer number from 0-63 (six binary bit).

Support for the indicator of Caps Lock and Numlock is removed, since I
do not depend on them for my workflow.